+++ services/user/CommonSys/common-lib/packages/wasm-loader/src/main.ts
@@ -1,9 +1,5 @@
 import { buildMessageLoaderInitialized } from "../../../../common/messaging/supervisor/LoaderInitialized";
-import { $init, provider } from "./component_parser";
-import {
-    generateFulfilledFunction,
-    generatePendingFunction
-} from "./dynamicFunctions";
+import { serviceMethodIndexToImportables } from "./dynamicFunctions";
 import importableCode from "./importables.js?raw";
 import {
     PluginCallResponse,
@@ -12,10 +8,11 @@ import {
     PluginCallPayload,
     buildPluginCallResponse
 } from "@messaging";
+import { parseFunctions } from "./witParsing";
 
 document.querySelector<HTMLDivElement>("#app")!.innerHTML = `
   <div>
-    <h1>Loader</h1>
+    <h1>WASM Loader</h1>
     <p>This is a generated SPA designed to act as the loader, its sole purpose is to be rendered in an iframe and run a WASM Component, then execute functions in the WASM component and send the results back to its parent iframe.</p>
   </div>
 `;
@@ -24,43 +21,25 @@ interface Importables {
     [key: string]: string;
 }
 
-const runWasm = async (
-    wasm: ArrayBuffer,
-    importables: Importables[],
-    method: string,
-    params: any[]
-) => {
-    console.info("runWasm().top 2");
-    const { load } = await import("rollup-plugin-wit-component");
-
-    console.time("Load");
-    const mod = await load(
-        // @ts-expect-error fff
-        wasm,
-        importables
-    );
-    console.timeEnd("Load");
-
-    return mod[method](...params);
-};
+const wasmUrlToIntArray = (url: string) =>
+    fetch(url)
+        .then((res) => res.arrayBuffer())
+        .then((buffer) => new Uint8Array(buffer));
 
 const functionCall = async ({
     id,
     args,
     precomputedResults
 }: PluginCallPayload) => {
-    const url = "/plugin.wasm";
-
-    const [wasmBytes] = await Promise.all([
-        fetch(url).then((res) => res.arrayBuffer()),
-        $init
+    const [pluginBytes, parserBytes, wasmComponent] = await Promise.all([
+        wasmUrlToIntArray("/plugin.wasm"),
+        wasmUrlToIntArray("/common/wasm-loader/component_parser.wasm"),
+        import("rollup-plugin-wit-component")
     ]);
-    const bytes = new Uint8Array(wasmBytes);
 
+    const { provider } = await wasmComponent.load(parserBytes);
     const ComponentParser = new provider.ComponentParser();
-
-    const { wit } = ComponentParser.parse("hello", bytes);
-
+    const { wit } = ComponentParser.parse("hello", pluginBytes);
     const parsedFunctions = parseFunctions(wit);
 
     const serviceImports = serviceMethodIndexToImportables(
@@ -81,12 +60,9 @@ const functionCall = async ({
     ];
 
     try {
-        const res = await runWasm(
-            wasmBytes,
-            importables,
-            args.method,
-            args.params
-        );
+        // @ts-ignore
+        const pluginModule = await wasmComponent.load(pluginBytes, importables);
+        const res = pluginModule[args.method](...args.params);
         console.log(`Plugin call ${args.service} Success: ${res}`);
         sendPluginCallResponse(buildPluginCallResponse(id, res));
     } catch (e) {
@@ -101,96 +77,6 @@ const sendPluginCallResponse = (response: PluginCallResponse) => {
 
 const onPluginCallRequest = (request: PluginCallRequest) =>
     functionCall(request.payload);
-
-const stripGenerated = (lines: string[]) => {
-    let startingIndex = -1;
-    let endingIndex = -1;
-
-    lines.forEach((line, index) => {
-        if (line.startsWith("package component:")) {
-            startingIndex = index;
-        } else {
-            if (startingIndex != -1 && endingIndex == -1) {
-                if (line.startsWith("package ")) {
-                    endingIndex = index;
-                }
-            }
-        }
-    });
-    return lines.slice(startingIndex, endingIndex);
-};
-
-const groupInterfaces = (lines: string[]) => {
-    let obj: { [key: string]: string[] } = {};
-    let isOpen = false;
-    let currentName = "";
-    lines.forEach((line) => {
-        if (line.startsWith("interface ")) {
-            isOpen = true;
-            currentName = line.split(" ")[1];
-        } else if (line == "}") {
-            isOpen = false;
-        } else if (isOpen) {
-            obj[currentName] = obj[currentName]
-                ? [...obj[currentName], line]
-                : [line];
-        }
-    });
-
-    return obj;
-};
-
-type ServiceMethodIndex = {
-    [k: string]: string[];
-};
-const parseFunctions = (data: string): ServiceMethodIndex => {
-    const lines = data.split("\n").map((x) => x.trim());
-    const newLines = stripGenerated(lines);
-
-    const interfaces = groupInterfaces(newLines);
-    const entries = Object.fromEntries(
-        Object.entries(interfaces).map(([key, value]): [string, string[]] => [
-            key,
-            value.map((v) => v.split(" ")[0].slice(0, -1))
-        ])
-    );
-
-    console.log({ interfaces, entries });
-
-    return entries;
-};
-
-interface FunctionResult<T = any> {
-    method: string;
-    result: T;
-}
-
-const serviceMethodIndexToImportables = (
-    serviceMethodIndex: ServiceMethodIndex,
-    service: string,
-    id: string,
-    functionsResult: FunctionResult[]
-): { [key: string]: string }[] =>
-    Object.entries(serviceMethodIndex).map(([key, methodNames]) => ({
-        [`component:${service}/${key}`]: methodNames
-            .map((methodName) => {
-                const functionResult = functionsResult.find(
-                    (funcResult) => funcResult.method == methodName
-                );
-                return (
-                    (functionResult
-                        ? generateFulfilledFunction(
-                              functionResult.method,
-                              functionResult.result
-                          )
-                        : generatePendingFunction(
-                              { method: methodName, service: key },
-                              id
-                          )) + "\n"
-                );
-            })
-            .join("\n")
-    }));
 
 const onRawEvent = (message: MessageEvent) => {
     if (isPluginCallRequest(message.data)) {
